Move all client map operations to go routine
Accessing Realm::clients from handleSessions is racey.
Initially I tried using Mutexes but was unable to find the correct
pattern to lock out the modification of the datastructure safely from
multiple threads, possibly this was related to not have barriers; really
don't know.

Using guidance from GOTPL I moved all the map operations to it's own go
routine and provided channels and functions which write to those
channels to control access to the map. This seems to solve a vast
majority of the data races reported by -race when running unit tests.