Add a hook to allow overriding of CheckOrigin
By default the websocket code verifies the origin of the caller
matches the url of the server and rejects the connection.
Gorilla provides a way to override this behavior. Provide a hook on
turnpike to allow providing an alternate function so the caller can
override the behavior. This is useful in a dev environment for front-end
work where the UI is being delivered from a node web server with
live-reload and the API server is on a different port.

- add hook, SetCheckOrigin(func (r *http.Request) bool), on
  turnpike.WebsocketServer which exposes gorilla websocket
  CheckOrigin so user of turnpike can disable same origin check.