Remove IsExpired hack
This is no longer necessary as we are enforcing authentication using
tickets in production.  This hack doesn't let the client recover gracefully

Adding a decorator to each procedure call is much cleaner and lets you
use the much cleaner error result which gives the clients more control
over how to retry the calls that fail.